refactor(modules): extract queue-related static methods from main module
